layout: gallery and news moved to template dependent html
now the html for the effects of these layered classes is handled in
Layout and loaded from the template file.
These classes just provide the JS functions and declare which ones they
provide.
In Layout is not yet implemented the ability to see if we are using a
not implemented function.

Still needs to be migrated the Hover handling to the Layout class.